Implementacion de procesos se crearon dos procesos el primer proceso es el menu del usuario el segundo proceso llama la funcion search y se queda esperando por datos

+++ leer.h
@@ -95,7 +95,7 @@ Nodo_L* buscar_binario(FILE* binario, int sourceid, int dstid, int hod) {
 
 
 
-int search(int sourceid, int dstid, int hod) {
+float search(int sourceid, int dstid, int hod) {
     // Abrir el archivo binario para lectura
     FILE* binario = fopen("binario.bin", "rb");
     if (binario == NULL) {
